ai-learning-dashboard/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
openai==1.30.1
Pillow==10.1.0
django-storages==1.14.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
whitenoise==6.6.0
